requirements.txt
streamlit
langchain
langchain-openai
langchain-community
PyPDF2
inspeqai
python-dotenv
# chromadb
# faiss-cpu
# tiktoken
# textstat
# evaluate
sentence_transformers
# tensorflow
# bert-score
# spacy
# nltk
# rouge_score
# sacrebleu
# sacremoses
# fuzzywuzzy
# python-Levenshtein
# levenshtein
# rank_bm25
tf-keras
lancedb
requests
beautifulsoup4